Method signature definition
The signature of a method consists of the name of the method and the type and kind (value, reference, or output) of each of its formal parameters, considered in the order left to right. The signature of a method specifically does NOT include the return type, nor does it include the params modifier that may be specified for the right-most parameter.